backend/foodgram/requirements.txt
Django==3.0.5
pytz==2021.1
sqlparse==0.4.1 
django-colorfield==0.4.2
Pillow==7.0.0
djoser==2.0.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.8.0
django-utils-six==2.0
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyJWT==2.1.0
python-dotenv==0.19.0
asgiref==3.2.10
django-environ==0.6.0
